+++ datafusion/physical-plan/src/joins/hash_join.rs
@@ -81,7 +81,7 @@ use datafusion_physical_expr::equivalence::{
 use datafusion_physical_expr::PhysicalExprRef;
 use datafusion_physical_expr_common::datum::compare_op_for_nested;
 
-use crate::poll_budget::{PollBudget, YieldStream};
+use crate::r#yield::{PollBudget, YieldStream};
 use ahash::RandomState;
 use datafusion_physical_expr_common::physical_expr::fmt_sql;
 use futures::{ready, Stream, StreamExt, TryStreamExt};
